requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
gradio>=5.0.0,<6.0.0
huggingface_hub>=0.24.0,<1.0.0
